backend/pyproject.toml
[project]
name = "account-platform-backend"
version = "1.0.0"
description = "Backend API for Account Platform - AWS account management system"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Account Platform Team", email = "team@example.com"}
]
keywords = ["fastapi", "aws", "account-management", "dynamodb", "kms"]

dependencies = [
    # Web Framework
    "fastapi>=0.119.0",
    "uvicorn[standard]>=0.37.0",
    "python-multipart>=0.0.9",

    # AWS SDK
    "boto3>=1.42.16",
    "botocore>=1.42.16",

    # Data Validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # Environment Management
    "python-dotenv>=1.0.0",

    # JWT validation (for Cognito)
    "python-jose[cryptography]>=3.3.0",

    # HTTP Client
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",

    # AWS Mocking
    "moto[all]>=5.0.0",

    # Code Quality
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",

    # Type Stubs
    "boto3-stubs[dynamodb,kms,sts,bedrock,servicequotas]>=1.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "-v",
    "--strict-markers",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "W", "F", "I", "C", "B", "UP"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
no_implicit_optional = true
